requirements.txt
langchain==0.1.20
langchain_community==0.0.38
langchain_core==0.1.52
langchain_openai==0.1.6
python-dotenv==1.0.1
youtube-search
neo4j